Moved authentication to routes
