+++ app/__init__.py
@@ -36,9 +36,9 @@ def create_app(**kwargs):
 
 def register_blueprint():
     from app.rest import base_blueprint
-    from app.authentication.rest import auth_blueprint
+    from app.routes.authentication.rest import auth_blueprint
     from app.routes.events.rest import events_blueprint
-    from app.fees.rest import fees_blueprint, fee_blueprint
+    from app.routes.fees.rest import fees_blueprint, fee_blueprint
     from app.routes.event_types.rest import event_types_blueprint, event_type_blueprint
     from app.routes.speakers.rest import speakers_blueprint, speaker_blueprint
     application.register_blueprint(base_blueprint)
