+++ tests/app/dao/test_blacklist_dao.py
@@ -6,7 +6,7 @@ from flask_jwt_extended import (
     create_refresh_token,
 )
 
-from app.authentication.errors import TokenNotFound
+from app.routes.authentication.errors import TokenNotFound
 from app.dao.blacklist_dao import store_token, is_token_revoked, get_user_tokens, unrevoke_token, prune_database
 from app.models import TokenBlacklist
 from tests.conftest import get_unixtime_start_and_expiry
